Change Skel library breakpoints
Implementing changes recommended by ChatGPT. 
https://twitter.com/awalkadayart/status/1619012171610075136

To display at least two vertical rows on a smartphone, I modified the small breakpoints to have a smaller max-width value. 
This will set the small breakpoints to be triggered when the screen width is less than 600px, which should display two or more rows on most smartphones.